Move G1PickAndPlaceMimicEnvCfg into pick_and_place_task.py
The cfg lived alone in g1_pick_and_place_mimic_cfg.py, the only file
in isaaclab_arena/tasks/ that contained no Task class. The repo
convention is "one file per task containing the task plus all its
related cfgs" (PickPlaceMimicEnvCfg lives next to PickAndPlaceTask,
PlaceUprightMimicEnvCfg lives next to PlaceUprightTask, etc.).

Move G1PickAndPlaceMimicEnvCfg next to PickPlaceMimicEnvCfg in
pick_and_place_task.py. Both cfgs are pick-and-place cfgs; having them
side by side makes the parent / G1-variant relationship visible and
positions the eventual follow-up (absorb the G1 shape into the parent)
as a single-file change.

Imports collapse from two lines to one in env, tests, and docs.

Behavior unchanged. Tests pass.
