--- isaaclab_arena/tasks/g1_pick_and_place_mimic_cfg.py
@@ -1,299 +1,0 @@
-# Copyright (c) 2025-2026, The Isaac Lab Arena Project Developers (https://github.com/isaac-sim/IsaacLab-Arena/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-from dataclasses import MISSING
-
-from isaaclab.envs.mimic_env_cfg import MimicEnvCfg, SubTaskConfig
-from isaaclab.utils import configclass
-
-from isaaclab_arena.embodiments.common.arm_mode import ArmMode
-
-
-@configclass
-class G1PickAndPlaceMimicEnvCfg(MimicEnvCfg):
-    """
-    Isaac Lab Mimic environment config class for the G1 humanoid pick-and-place tasks.
-
-    Encodes G1-specific cfg shape: 3 subtasks per arm
-    (idle / grasp_and_idle / final), a 4-phase nav body subtask sequence
-    (navigate_to_table / navigate_turn_inplace / navigate_to_bin / final), and
-    G1-specific per-subtask numerical tunings. Used by the locomanip env via
-    ``PickAndPlaceTask(mimic_env_cfg_factory=...)``.
-    """
-
-    pick_up_object_name: str = MISSING
-    destination_name: str = MISSING
-    arm_mode: ArmMode = ArmMode.DUAL_ARM
-
-    def __post_init__(self):
-        # post init of parents
-        super().__post_init__()
-
-        # Hardcoded right + left + body subtask shapes assume both arms drive the demo;
-        # other arm modes would silently produce a cfg with the wrong subtask shape.
-        if self.arm_mode != ArmMode.DUAL_ARM:
-            raise ValueError(f"G1PickAndPlaceMimicEnvCfg only supports ArmMode.DUAL_ARM; got {self.arm_mode}")
-
-        self.datagen_config.name = f"locomanip_pick_and_place_{self.pick_up_object_name}_to_{self.destination_name}_D0"
-        self.datagen_config.generation_guarantee = True
-        self.datagen_config.generation_keep_failed = False
-        self.datagen_config.generation_num_trials = 100
-        self.datagen_config.generation_select_src_per_subtask = False
-        self.datagen_config.generation_select_src_per_arm = False
-        self.datagen_config.generation_relative = False
-        self.datagen_config.generation_joint_pos = False
-        self.datagen_config.generation_transform_first_robot_pose = False
-        self.datagen_config.generation_interpolate_from_last_target_pose = True
-        self.datagen_config.max_num_failures = 25
-        self.datagen_config.seed = 1
-
-        # Right arm subtasks
-        subtask_configs = []
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                subtask_term_signal="idle_right",
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.003,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                subtask_term_signal="grasp_and_idle_right",
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.003,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.003,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        self.subtask_configs["right"] = subtask_configs
-
-        # Left arm subtasks
-        subtask_configs = []
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                subtask_term_signal="idle_left",
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.003,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                subtask_term_signal="grasp_and_idle_left",
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.003,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.003,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        self.subtask_configs["left"] = subtask_configs
-
-        # Body subtasks (used for navigation)
-        subtask_configs = []
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                subtask_term_signal="navigate_to_table",
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.0,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                subtask_term_signal="navigate_turn_inplace",
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.0,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                subtask_term_signal="navigate_to_bin",
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.0,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        subtask_configs.append(
-            SubTaskConfig(
-                # Each subtask involves manipulation with respect to a single object frame.
-                object_ref=self.pick_up_object_name,
-                # This key corresponds to the binary indicator in "datagen_info" that signals
-                # when this subtask is finished (e.g., on a 0 to 1 edge).
-                first_subtask_start_offset_range=(0, 0),
-                # Randomization range for starting index of the first subtask
-                subtask_term_offset_range=(0, 0),
-                # Selection strategy for the source subtask segment during data generation
-                selection_strategy="nearest_neighbor_object",
-                # Optional parameters for the selection strategy function
-                selection_strategy_kwargs={"nn_k": 3},
-                # Amount of action noise to apply during this subtask
-                action_noise=0.0,
-                # Number of interpolation steps to bridge to this subtask segment
-                num_interpolation_steps=0,
-                # Additional fixed steps for the robot to reach the necessary pose
-                num_fixed_steps=0,
-                # If True, apply action noise during the interpolation phase and execution
-                apply_noise_during_interpolation=False,
-            )
-        )
-        self.subtask_configs["body"] = subtask_configs
